Move versioning test helpers to test environment
